% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inhomPP_piecewiseconst}
\alias{inhomPP_piecewiseconst}
\title{Sample an Inhomogeneous Poisson Process with Piecewise Constant Intensity}
\usage{
inhomPP_piecewiseconst(tvec, lambdavec, tmax, first)
}
\arguments{
\item{tvec}{times at which the intensity changes (in j=0,1,...,k)}

\item{lambdavec}{values at the left endpoints of intensity (in j=0,1,...,k)}

\item{tmax}{maximum simulation time (usually equal to value of \code{tvec} at k)}

\item{first}{if \code{TRUE}, stop early and return the first event in the process}
}
\value{
a vector of event times in the counting process
}
\description{
Sample an inhomogeneous Poisson process withe piecewise constant intensity using an algorithm presented in:
\itemize{
\item Leemis, Lawrence M., and Stephen Keith Park. Discrete-event simulation: A first course. Upper Saddle River, NJ: Pearson Prentice Hall, 2006.
}
}
