% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inhomPP_piecewiseconst_reject}
\alias{inhomPP_piecewiseconst_reject}
\title{Sample First Event Time Inhomogeneous Poisson Process with Piecewise Constant Intensity}
\usage{
inhomPP_piecewiseconst_reject(tvec, lambdavec)
}
\arguments{
\item{tvec}{times at which the intensity changes (in j=0,1,...,k)}

\item{lambdavec}{values at the left endpoints of intensity (in j=0,1,...,k)}
}
\value{
a vector of number of rejections and the time of first event
}
\description{
Sample an inhomogeneous Poisson process withe piecewise constant intensity using an accept-reject algorithm.
If the first event occurs after the time that the intensity is defined via \code{lambdavec}, return a \code{NaN}
}
