% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR-finalsize.R
\name{SIR_finalsize}
\alias{SIR_finalsize}
\title{Compute Epidemic Final Size Distribution for SIR Model}
\usage{
SIR_finalsize(N, m, lambda, phi)
}
\arguments{
\item{N}{initial numer of susceptibles}

\item{m}{initial number of infectives}

\item{lambda}{the effective contact rate for a force of infection term following the mass action law}

\item{phi}{the MGF of the infectious period distribution (see \url{https://en.wikipedia.org/wiki/Moment-generating_function#Examples} for MGFs of various named distributions)}
}
\description{
Compute the final epidemic size of a stochastic SIR epidemic with arbitrary distribution of infectious period using the closed form method of:
\itemize{
\item Ball, Frank. "A unified approach to the distribution of total size and total area under the trajectory of infectives in epidemic models." Advances in Applied Probability 18.2 (1986): 289-310.
}
}
\details{
This method becomes numerically unstable for N larger than about 100.
}
\examples{
\dontrun{
# this reproduces figure 2.2 from the book: Andersson, Hakan, and Tom Britton. Stochastic epidemic models and their statistical analysis. Vol. 151. Springer Science & Business Media, 2012.
# N=50 susceptibles and m=1 infected at time 0
 N <- 50
 lambda <- 1.5
 m <- 1
# duration of infectiousness is deterministic and lasts 1 day
 phi <- function(b){
   exp(b)
 }

 probs <- SIR_finalsize(N = N,m = m,lambda = lambda,phi = phi)
 plot(0:N,probs,pch=16)
 abline(h=0)
}

}
