% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SIRMarkov_ABM}
\alias{SIRMarkov_ABM}
\title{Simulate Markovian SIR Model via Agent-based Model (ABM)}
\usage{
SIRMarkov_ABM(dt, tmax, S, I, R, beta, gamma, verbose)
}
\arguments{
\item{dt}{the time step}

\item{tmax}{maximum time of simulation}

\item{S}{initial number of susceptible individuals}

\item{I}{initial number of infected & infectious individuals}

\item{R}{initial number of recovered individuals}

\item{beta}{the product of transmission probability and contact rate}

\item{gamma}{inverse of the duration of infectiousness}

\item{verbose}{print extra information?}
}
\value{
a list (use \code{do.call(cbind,out)} to convert to \code{matrix})
}
\description{
Sample a trajectory from the Markovian SIR model using the approximate ABM.
}
