% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SIRMarkov_MNRM}
\alias{SIRMarkov_MNRM}
\title{Simulate Markovian SIR Model via Modified Next Reaction Method (MNRM)}
\usage{
SIRMarkov_MNRM(tmax, S, I, R, beta, gamma, verbose)
}
\arguments{
\item{tmax}{maximum time of simulation}

\item{S}{initial number of susceptible individuals}

\item{I}{initial number of infected & infectious individuals}

\item{R}{initial number of recovered individuals}

\item{beta}{the product of transmission probability and contact rate}

\item{gamma}{inverse of the duration of infectiousness}

\item{verbose}{print extra information?}
}
\value{
a matrix
}
\description{
Sample a trajectory from the Markovian SIR model using the MNRM algorithm presented in:
\itemize{
\item Anderson, D. F. (2007). A modified next reaction method for simulating chemical systems with time dependent propensities and delays. Journal of Chemical Physics, 127(21). \url{https://doi.org/10.1063/1.2799998}
}
}
