% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocheulerABM.R
\docType{package}
\name{stocheulerABM-package}
\alias{stocheulerABM-package}
\title{stocheulerABM: a cool package}
\description{
This R package was developed by
Sean L. Wu (slwu89@berkeley.edu)
in the Divisions of Biostatistics and Epidemiology at UC Berkeley, California, USA.
}
\section{stocheulerABM functions}{

}

